Added BE for Portfolio manager

+++ users/views.py
@@ -11,7 +11,7 @@ from django.views.decorators.csrf import csrf_exempt
 
 from .forms import LoginForm
 from .forms import RegisterForm
-from .models import WatchListItem
+from .models import WatchListItem, PortfolioItem
 
 
 # Create your views here.
@@ -51,14 +51,13 @@ def add_to_watchlist(request):
         name = request_data['name']
         symbol = request_data['symbol']
         print(name, symbol)
-        if name and symbol:
-            try:
-                WatchListItem.objects.create(user=request.user, name=name, symbol=symbol)
-                return JsonResponse(dict(success=True, msg=f"{name} added to watchlist"), status=201)
-            except IntegrityError:
-                return JsonResponse({'success': False, 'message': 'Item already in watchlist'}, status=400)
-        else:
+        if not name or not symbol:
             return JsonResponse(dict(success=False, msg="Invalid data"), status=400)
+        try:
+            WatchListItem.objects.create(user=request.user, name=name, symbol=symbol)
+            return JsonResponse(dict(success=True, msg=f"{name} added to watchlist"), status=201)
+        except IntegrityError:
+            return JsonResponse({'success': False, 'message': 'Item already in watchlist'}, status=400)
     else:
         return JsonResponse(dict(success=False, msg="Invalid request method"), status=405)
 
@@ -77,13 +76,83 @@ def delete_from_watchlist(request):
         request_data = json.loads(request.body)
         print(request_data)
         symbol = request_data['symbol']
-        if symbol:
-            WatchListItem.objects.filter(
+        if not symbol:
+            return JsonResponse(dict(success=False, msg="Invalid data"), status=400)
+        WatchListItem.objects.filter(
+            symbol=symbol,
+            user=request.user
+        ).delete()
+        return JsonResponse(dict(success=True, msg=f"{symbol} removed from watchlist"), status=201)
+    else:
+        return JsonResponse(dict(success=False, msg="Invalid request method"), status=405)
+
+
+@csrf_exempt
+@login_required
+def add_to_portfolio(request):
+    if request.method == "POST":
+        request_data = json.loads(request.body)
+        print(request_data)
+        name = request_data['name']
+        symbol = request_data['symbol']
+        amount = request_data['amount']
+        print(name, symbol)
+        if not name or not symbol or amount <= 0:
+            return JsonResponse(dict(success=False, msg="Invalid data"), status=400)
+        try:
+            PortfolioItem.objects.create(user=request.user, name=name, symbol=symbol, quantity=amount)
+            return JsonResponse(dict(success=True, msg=f"{amount} {symbol} added to portfolio"), status=201)
+        except IntegrityError:
+            return JsonResponse({'success': False, 'message': 'Item already in watchlist'}, status=400)
+    else:
+        return JsonResponse(dict(success=False, msg="Invalid request method"), status=405)
+
+@csrf_exempt
+@login_required
+def update_amount_portfolio(request):
+    if request.method == "POST":
+        request_data = json.loads(request.body)
+        symbol = request_data['symbol']
+        amount = request_data['amount']
+        print(symbol, amount)
+        if not symbol or amount < 0:
+            return JsonResponse(dict(success=False, msg="Invalid data"), status=400)
+        if amount == 0:
+            PortfolioItem.objects.filter(symbol=symbol, user=request.user).delete()
+            return JsonResponse(dict(success=True, msg=f"{symbol} removed from portfolio"), status=200)
+        try:
+            PortfolioItem.objects.filter(
                 symbol=symbol,
                 user=request.user
-            ).delete()
-            return JsonResponse(dict(success=True, msg=f"{symbol} removed from watchlist"), status=201)
+            ).update(amount=amount)
+            return JsonResponse(dict(success=True, msg=f"{symbol} amount change {amount}"), status=201)
+        except IntegrityError:
+            return JsonResponse({'success': False, 'message': 'Integrity Error'}, status=400)
+    else:
+        return JsonResponse(dict(success=False, msg="Invalid request method"), status=405)
+
+
+def delete_from_portfolio(request):
+    if request.method == "POST":
+        request_data = json.loads(request.body)
+        symbol = request_data['symbol']
+        if symbol:
+            try:
+                PortfolioItem.objects.filter(
+                    symbol=symbol,
+                    user=request.user
+                ).delete()
+                return JsonResponse(dict(success=True, msg=f"{symbol} removed from portfolio"), status=201)
+            except IntegrityError:
+                return JsonResponse({'success': False, 'message': 'Integrity Error'}, status=400)
         else:
             return JsonResponse(dict(success=False, msg="Invalid data"), status=400)
     else:
         return JsonResponse(dict(success=False, msg="Invalid request method"), status=405)
+
+
+@login_required
+def get_portfolio(request):
+    portfolio_items = WatchListItem.objects.filter(user=request.user)
+    data = [{"name": item.name, "symbol": item.symbol, "amount": item.amount} for item in portfolio_items]
+    return JsonResponse(data, safe=False)
